NonlinearCrystals

Documentation for NonlinearCrystals.

NonlinearCrystals.Angle

NonlinearCrystals.AngleFreeUnits

NonlinearCrystals.AngleUnits

NonlinearCrystals.contract_voigt_index

NonlinearCrystals.expand_voigt_index

NonlinearCrystals.rot_mat_crys_to_diel

NonlinearCrystals.Angle — Type

NonlinearCrystals.Angle{T, U}

A supertype for quantities and levels of dimension Unitful.𝐋 ^ 0 with a value of type T and units U.

See also: Unitful.Quantity, Unitful.Level.

NonlinearCrystals.AngleFreeUnits — Type

NonlinearCrystals.AngleFreeUnits{U}

A supertype for Unitful.FreeUnits of dimension Unitful.𝐋 ^ 0. Equivalent to Unitful.FreeUnits{U, Unitful.𝐋 ^ 0}.

NonlinearCrystals.AngleUnits — Type

NonlinearCrystals.AngleUnits{U}

A supertype for units of dimension Unitful.𝐋 ^ 0. Equivalent to Unitful.Units{U, Unitful.𝐋 ^ 0}.

See also: Unitful.Units.

NonlinearCrystals.contract_voigt_index — Method

contract_voigt_index(i::Integer, j::Integer, k::Integer) -> (i::Int, l::Int)

Given full tensor indices (i, j, k), return the contracted Voigt index (i, l). This assumes symmetry in (j, k): for example, (1,3) and (3,1) both map to l = 5.

NonlinearCrystals.expand_voigt_index — Method

expand_voigt_index(i::Integer, l::Integer) -> Vector{NTuple{3,Int}}

Given contracted Voigt indices (i, l), return list of (i, j, k) index tuples. Some l values map to two symmetric (j, k) combinations.

NonlinearCrystals.rot_mat_crys_to_diel — Method

rot_mat_crys_to_diel(
    axes_assignment_crys_to_diel::NTuple{3,Symbol};
    rotate_about::Union{Symbol,Nothing}=nothing,
    phi::Angle=0.0u"°"
)

Returns the rotation matrix from crystallophysical (x,y,z) to dielectric (X,Y,Z) coordinates, by optionally rotating around a dielectric axis after reassigning the axes.

axes_assignment_crys_to_diel: 3-Tuple of Symbols (:X, :Y, :Z) indicating the assignment of crystallophysical to dielectric coordinates

rotate_about: Symbol (:X, :Y, or :Z) indicating which dielectric axis to rotate around after assignment

phi: rotation angle (applied after assignment)